roleready/apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-docx==1.1.2
pdfminer.six==20240706
scikit-learn==1.5.2
nltk==3.9.1
pymupdf==1.24.9           # detect text-layer + fast PDF text
pytesseract==0.3.13       # OCR (local)
pdf2image==1.17.0         # render PDF pages to images for OCR
Pillow==10.4.0            # imaging backend for OCR
sentence-transformers==2.7.0
torch==2.3.1
numpy==1.26.4
openai==1.43.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
supabase==2.5.0
docxtpl==0.16.7
jinja2==3.1.4
python-jose[cryptography]==3.3.0
requests==2.32.3
